requirements.txt
discord.py
apscheduler
gspread
google-auth